♻️ Moving duplicate code to common
