+++ components/pages/submission-system/program-dashboard/DonorDataSummary/DonorSummaryTable.tsx
@@ -12,6 +12,7 @@ import {
   DataTableStarIcon as StarIcon,
   TableInfoHeaderContainer,
   CellContentCenter,
+  Pipeline,
 } from '../../common';
 
 import { createRef } from 'react';
@@ -20,17 +21,9 @@ import { useTheme } from 'uikit/ThemeProvider';
 import { css } from '@emotion/core';
 import styled from '@emotion/styled';
 import HyperLink from 'uikit/Link';
-import Pipe from 'uikit/Pipe';
 import DonorStatsArea from './DonorSummaryTableStatArea';
 import { RELEASED_STATE_FILL_COLOURS, RELEASED_STATE_STROKE_COLOURS } from './common';
 import { startCase } from 'lodash';
-
-enum PIPELINE_STATUS {
-  COMPLETE = 'complete',
-  IN_PROGRESS = 'inProgress',
-  ERROR = 'error',
-}
-type PipelineStats = Record<PIPELINE_STATUS, number>;
 
 export default ({
   donorSummaryEntries,
@@ -141,33 +134,6 @@ export default ({
     );
   };
 
-  const PipelineCell = (stats: PipelineStats) => {
-    const theme = useTheme();
-
-    const getBackgroundColour = (state: keyof PipelineStats) => {
-      interface ColourMapper {
-        [key: string]: keyof typeof theme.colors;
-      }
-      const mapper: ColourMapper = {
-        [PIPELINE_STATUS.COMPLETE]: 'accent1_dimmed',
-        [PIPELINE_STATUS.IN_PROGRESS]: 'warning_dark',
-        [PIPELINE_STATUS.ERROR]: 'error',
-      };
-      return mapper[state];
-    };
-
-    const shouldRender = (num: number) => num > 0;
-
-    const renderableStats = Object.keys(stats).filter(key => shouldRender(stats[key]));
-
-    const pipeStats = renderableStats.map(stat => (
-      <Pipe.Item key={stat} fill={getBackgroundColour(stat as keyof PipelineStats)}>
-        {stats[stat]}
-      </Pipe.Item>
-    ));
-    return <Pipe>{pipeStats}</Pipe>;
-  };
-
   const tableColumns: Array<TableColumnConfig<DonorSummaryRecord>> = [
     {
       Header: 'CLINICAL DATA STATUS',
@@ -253,7 +219,7 @@ export default ({
           Header: 'Alignment',
           accessor: 'alignment',
           Cell: ({ original }) => (
-            <PipelineCell
+            <Pipeline
               complete={original.alignmentsCompleted}
               inProgress={original.alignmentsRunning}
               error={original.alignmentsFailed}
@@ -264,7 +230,7 @@ export default ({
           Header: 'Sanger VC',
           accessor: 'sangerVC',
           Cell: ({ original }) => (
-            <PipelineCell
+            <Pipeline
               complete={original.sangerVcsCompleted}
               inProgress={original.sangerVcsRunning}
               error={original.sangerVcsFailed}
